chore(code-health-rules): Move rules-code to its own module
